--appropriately initialize maps for expected default vals.

+++ src/esp/metadata/attributes/ObjectAttributes.h
@@ -299,7 +299,9 @@ class ObjectAttributes : public AbstractObjectAttributes {
   void setIsVisible(bool isVisible) { set("is_visible", isVisible); }
   bool getIsVisible() const { return getBool("is_visible"); }
 
-  void setSemanticId(uint32_t semanticId) { set("semantic_id", semanticId); }
+  void setSemanticId(uint32_t semanticId) {
+    set("semantic_id", (int)semanticId);
+  }
 
   uint32_t getSemanticId() const { return getInt("semantic_id"); }
